requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Machine Learning
tensorflow==2.13.0
numpy==1.24.3
pillow==10.0.1

# Firebase
firebase-admin==6.2.0

# Data Processing
pandas==2.0.3
scikit-learn==1.3.0

# Environment
python-dotenv==1.0.0